fix: break circular dependencies in ai-elements component splits
- Move AttachmentsContext type to prompt-input-attachments-context.ts
  to break cycle with prompt-input-context.tsx
- Extract ReasoningContext + useReasoning to reasoning-context.ts
  to break cycle between reasoning.tsx and reasoning-trigger.tsx
