+++ frontend/components/ai-elements/prompt-input-attachments-context.ts
@@ -5,8 +5,18 @@
  * Refresh can treat that file as a components-only module.
  */
 
-import { createContext } from 'react';
-import type { AttachmentsContext } from './prompt-input-context';
+import type { FileUIPart } from 'ai';
+import { createContext, type RefObject } from 'react';
+
+/** Attachment controller exposed to prompt input child components. */
+export type AttachmentsContext = {
+	files: (FileUIPart & { id: string })[];
+	add: (files: File[] | FileList) => void;
+	remove: (id: string) => void;
+	clear: () => void;
+	openFileDialog: () => void;
+	fileInputRef: RefObject<HTMLInputElement | null>;
+};
 
 /** Local (per-PromptInput) attachments context. */
 export const LocalAttachmentsContext = createContext<AttachmentsContext | null>(null);
